refactor(kinds): move the resource-kind tables out of ResourceBrowser
lib/paletteActions imported a UI component to read a lookup table, which is a
dependency pointing the wrong way and blocks the service layer being extracted.

ResourceKind, RESOURCE_LABELS and K8S_KIND are pure data and now live in
lib/kinds.ts. ResourceBrowser re-exports them so its other importers are
untouched for now; the next commit repoints them and drops the re-export.
